python-file-upload/requirements.txt
Flask==3.0.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
Werkzeug==3.0.1
gunicorn==21.2.0
